feat: postcard done, pause menu done!
