--- CookingStudent/Assets/Resources/PerformanceTestRunInfo.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5cc46e9d268cf8d4ba78f3ef838560eb
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
