--- CookingStudent/Assets/Resources/PerformanceTestRunSettings.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: df5b9356db6749e4c941c93bdfa606ca
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
