--- CookingStudent/Assets/Scripts/PauseUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 78acbd95eeb131b4fa8eb0425ddcea2c
