--- CookingStudent/Assets/Scripts/PostcardUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 41aa555438ef5f64a84fa4fed203fff7
