+++ CookingStudent/Assets/UI Elements/button.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d58e0d0745d48894b87f0533b8264070
+guid: df7d80dce6972594c886020939863081
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
